requirements.txt
# 核心依赖
python>=3.8

# Web框架
flask>=2.3.0
flask-cors>=4.0.0

# AI/LLM
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.2
mcp>=0.1.0

# 网络请求
requests>=2.31.0
aiohttp>=3.8.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 其他
colorama>=0.4.0
rich>=13.0.0
